added stuff to make snake find food
Snake will search for food when health gets to 50. Snake will go to the closest food piece to its head, however at this point it may collide with itself in doing so. Had to add in a this.test variable so I could access the request object.

Also added a gitignore that ignores node_modules folder